% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{counts}
\alias{counts}
\alias{counts,RmmquantClass-method}
\title{Get the counts table of an \code{RmmquantClass} object.}
\usage{
counts(object)

\S4method{counts}{RmmquantClass}(object)
}
\arguments{
\item{object}{An \code{RmmquantClass} object.}
}
\value{
The count matrix, in a \code{SummarizedExperiment}
}
\description{
Get the counts table of an \code{RmmquantClass} object.
}
\examples{
example <- RmmquantClassExample()
counts(example)

}
