% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmmquant.R
\docType{package}
\name{Rmmquant}
\alias{Rmmquant}
\title{Rmmquant: RNA-Seq multi-mapping Reads Quantification Tool}
\description{
Counts the number of reads per gene.
}
\author{
Matthias Zytnicki, \email{matthias.zytnicki@inra.fr}
}
