% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{RmmquantClass-class}
\alias{RmmquantClass-class}
\title{An S4 class for Rmmquant.}
\description{
An S4 class for Rmmquant.
}
\section{Slots}{

\describe{
\item{\code{annotationFile}}{The annotation file}

\item{\code{readsFiles}}{The reads files}

\item{\code{genomicRanges}}{The annotation, in a
\code{GenomicRanges} format.}

\item{\code{genomicRangesList}}{The annotation, in a
\code{GenomicRangesList} format.}

\item{\code{sampleNames}}{The name of the samples}

\item{\code{overlap}}{The minimum number of overlapping base pairs to
declare a match.}

\item{\code{strands}}{Whether annotation of the same strand should be
considered.}

\item{\code{sorts}}{Whether the files are sorted.}

\item{\code{countThreshold}}{The reads files}

\item{\code{mergeThreshold}}{The reads files}

\item{\code{printGeneName}}{Whether the (vernacular) gene name is reported.}

\item{\code{quiet}}{Shut Rmmquant up.}

\item{\code{progress}}{Print the progress of the tool.}

\item{\code{nThreads}}{The number of threads.}

\item{\code{formats}}{The format of the reads files (SAM or BAM).}

\item{\code{nOverlapDiff}}{Difference of overlap between a primary map and a
secondary map.}

\item{\code{pcOverlapDiff}}{Ratio of overlap between a primary map and a
secondary map.}

\item{\code{counts}}{A matrix of the counts.}

\item{\code{stats}}{A data frame of the quantification statistics.}
}}

