% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{RmmquantClassExample}
\alias{RmmquantClassExample}
\title{Example of Rmmquant constructor.}
\usage{
RmmquantClassExample()
}
\value{
An \code{RmmquantClass}.
}
\description{
Example of Rmmquant constructor.
}
\examples{
example <- RmmquantExample()

}
