% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{RmmquantExample}
\alias{RmmquantExample}
\title{Example of Rmmquant use}
\usage{
RmmquantExample()
}
\value{
An \code{SummarizedExperiement}.
}
\description{
Example of Rmmquant use
}
\examples{
example <- RmmquantExample()

}
