% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{stats}
\alias{stats}
\alias{stats,RmmquantClass-method}
\title{Get the stats}
\usage{
stats(object)

\S4method{stats}{RmmquantClass}(object)
}
\arguments{
\item{object}{An \code{RmmquantClass} object.}
}
\value{
The stats data frame
}
\description{
Get the stats
}
\examples{
example <- RmmquantExample()
stats(example)

}
