% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{validateRmmquant}
\alias{validateRmmquant}
\title{Rmmquant object validation function}
\usage{
validateRmmquant(object)
}
\arguments{
\item{object}{A \code{RmmquantClass} object.}
}
\value{
\code{TRUE}, if succeed, otherwise a \code{character}.
}
\description{
Rmmquant object validation function
}
