% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,RmmquantClass-method}
\title{Get the counts}
\usage{
counts(object)

\S4method{counts}{RmmquantClass}(object)
}
\arguments{
\item{object}{An \code{RmmquantClass} object.}
}
\value{
The count matrix
}
\description{
Get the counts
}
\examples{
example <- RmmquantExample()
counts(example)

}
