% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{RmmquantExample}
\alias{RmmquantExample}
\title{Example constructor}
\usage{
RmmquantExample()
}
\value{
An \code{RmmquantClass} object
}
\description{
Example constructor
}
\examples{
example <- RmmquantExample()

}
