% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{show,RmmquantClass-method}
\alias{show,RmmquantClass-method}
\title{Overloading the show method}
\usage{
\S4method{show}{RmmquantClass}(object)
}
\arguments{
\item{object}{An \code{RmmquantClass} object.}
}
\value{
A description of the object.
}
\description{
Overloading the show method
}
\examples{
example <- RmmquantClassExample()
example

}
